refactor: split trash implementation for mac and linux
